+++ sources/views/Minifig/FigsListView.swift
@@ -94,14 +94,13 @@ struct MinifigListView: View {
             
             ForEach(values) { value in
                 NavigationLink(destination: MinifigDetailView(minifig: value)){
-                    FigsGridCell(minifig: value).id(value.id)
-                    
-                }.id(value.id)
+                    FigsGridCell(minifig: value)
+                }
                 .contentShape(RoundedRectangle(cornerRadius: 16, style: .continuous))
                 .padding(16)
                 .contextMenu{
                     contextMenuContent(value)
-                }.id(value.id)
+                }
                 
             }
         }